fix: comprehensive codebase audit and lint/test fixes
Full codebase audit across Go backend and React frontend, plus targeted
fixes for all lint errors and a data race:

Audit findings addressed (Go backend):
- DB layer: improved error handling, query consistency, migration safety
- API infrastructure: enhanced middleware, rate limiting, auth handlers
- Agent core/transport/apply: robust error propagation, context handling
- Rule engine: compiler/resolver edge cases, test coverage
- Alert system: evaluator robustness, SMTP formatting, spike detection
- Import pipeline: applier/resolver error paths
- Shared infra: metrics, log cleanup, context utilities

Audit findings addressed (React frontend):
- Components: MultiSelect, SearchableSelect, Toast, ConfirmModal improvements
- State/hooks: usePagination, useSSE, useToast, ToastContext refinements
- API client: error handling and request consistency
- Utilities: apiErrors formatting and test coverage

Lint and race fixes:
- Fix 3 Go errcheck errors (unchecked Close() in applier.go, shipper.go)
- Fix 2 Go misspell errors (cancelling→canceling, cancelled→canceled)
- Fix React Compiler memoization conflict in MultiSelect.jsx
- Fix data race in events/hub.go (RLock held through channel sends)
- Fix DB schema vs API handler spelling mismatch across migrations/schema/openapi
- Fix test comment spelling (cancelled→canceled) in 4 test files
- Fix stylistic inconsistency in SearchableSelect.jsx (useCallback removal)

Verification: make lint-go (0 issues), make lint-web (0 issues),
go test -race (no races detected).

+++ internal/api/settings/handlers.go
@@ -36,7 +36,7 @@ type LogSettings struct {
 	RetentionDays   int    `json:"retention_days"`
 	RetentionLabel  string `json:"retention_label"`
 	LogCount        int    `json:"log_count"`
-	EstimatedSizeMB int    `json:"estimated_size_mb"`
+	EstimatedSizeMB int64  `json:"estimated_size_mb"`
 	LogsDBPath      string `json:"logs_db_path"`
 }
 
@@ -56,7 +56,7 @@ func (h *Handler) GetLogSettings(w http.ResponseWriter, r *http.Request) {
 		logCount = 0
 	}
 
-	estimatedSizeMB := (logCount * 500) / (1024 * 1024)
+	estimatedSizeMB := int64(float64(logCount) * 500.0 / (1024.0 * 1024.0))
 
 	common.RespondJSON(w, http.StatusOK, LogSettings{
 		RetentionDays:   retentionDays,
